fix: satisfy workflow gui lint
